requirements.txt
timm==0.4.12
numpy==1.24.4
argparse
pyaml
pillow==10.2.0
opencv-python==4.9.0.80
tqdm
faiss-gpu==1.7.2
torch==2.1.2
torchvision==0.16.2
six
